latest commit

+++ work/prj1-sol/bits-to-ints.h
@@ -43,7 +43,7 @@ typedef unsigned long long BitsValue;
  */
 int getbits(FILE *inFile,bool *isEof);
 
-int getByte(FILE *inFile,bool *isEof);
+long long getByte(FILE *inFile,bool *isEof);
 BitsValue bits_to_ints(FILE *inFile, const char *inName, int nBits,
                        bool *isEof);
